# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Jobswyft API
  version: 1.0.0
  description: AI-powered job application assistant API
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.jobswyft.com
    description: Production
paths:
...
  /v1/resumes:
    get:
      summary: List user's resumes
      operationId: listResumes
      tags:
        - Resumes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paginated list of resumes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeListResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Upload and parse a resume
      operationId: uploadResume
      tags:
        - Resumes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file (max 10MB)
              required:
                - file
      responses:
        '200':
          description: Resume uploaded and parsed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeUploadResponse'
        '400':
          description: Validation error (not PDF or over 10MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Credit exhausted or resume limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/resumes/{resume_id}:
    get:
      summary: Get resume details with download URL
      operationId: getResume
      tags:
        - Resumes
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: resume_id
          schema:
            type: string
            format: uuid
          required: true
          description: Resume UUID
      responses:
        '200':
          description: Resume details with signed download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resume not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a resume
      operationId: deleteResume
      tags:
        - Resumes
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: resume_id
          schema:
            type: string
            format: uuid
          required: true
          description: Resume UUID
      responses:
        '200':
          description: Resume deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDeleteResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resume not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/resumes/{resume_id}/active:
    put:
      summary: Set resume as active
      operationId: setActiveResume
      tags:
        - Resumes
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: resume_id
          schema:
            type: string
            format: uuid
          required: true
          description: Resume UUID to set as active
      responses:
        '200':
          description: Resume set as active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetActiveResumeResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resume not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: AUTH_REQUIRED
            message:
              type: string
            details:
              type: object
...
    ResumeListItem:
      type: object
      required:
        - id
        - file_name
        - is_active
        - parse_status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        file_name:
          type: string
          example: john_doe_resume.pdf
        is_active:
          type: boolean
        parse_status:
          type: string
          enum: [pending, completed, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ResumeListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ResumeListItem'
                total:
                  type: integer
                page:
                  type: integer
                page_size:
                  type: integer
    ParsedResumeData:
      type: object
      properties:
        contact:
          type: object
          properties:
            first_name:
              type: string
            last_name:
              type: string
            email:
              type: string
            phone:
              type: string
            location:
              type: string
            linkedin_url:
              type: string
        summary:
          type: string
        experience:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              company:
                type: string
              start_date:
                type: string
              end_date:
                type: string
              description:
                type: string
        education:
          type: array
          items:
            type: object
            properties:
              degree:
                type: string
              institution:
                type: string
              graduation_year:
                type: string
        skills:
          type: array
          items:
            type: string
    ResumeDetailResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - id
                - file_name
                - file_path
                - is_active
                - parse_status
                - download_url
                - created_at
                - updated_at
              properties:
                id:
                  type: string
                  format: uuid
                file_name:
                  type: string
                file_path:
                  type: string
                is_active:
                  type: boolean
                parse_status:
                  type: string
                  enum: [pending, completed, failed]
                parsed_data:
                  $ref: '#/components/schemas/ParsedResumeData'
                download_url:
                  type: string
                  format: uri
                  description: Signed URL for PDF download (expires in 1 hour)
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
    ResumeUploadResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                resume:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    user_id:
                      type: string
                      format: uuid
                    file_name:
                      type: string
                    file_path:
                      type: string
                    parsed_data:
                      $ref: '#/components/schemas/ParsedResumeData'
                    parse_status:
                      type: string
                      enum: [pending, completed, failed]
                    created_at:
                      type: string
                      format: date-time
                    updated_at:
                      type: string
                      format: date-time
                ai_provider_used:
                  type: string
                  nullable: true
                  example: claude
    SetActiveResumeResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
                  example: Resume set as active
                active_resume_id:
                  type: string
                  format: uuid
    ResumeDeleteResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
                  example: Resume deleted successfully
...